Fixed use-after-free bug.
This arose when a paragraph containing only reference links and
blank space was finalized.  Finalization would remove the
node.  `finalize` returns the parent node, but the problem
arose because we had both `cur` and `parser->current`, and
only one was being updated.  Solution:  remove `cur`, which is
a holdover from before we had `parser->current`.

I believe this will close #9 -- @JordanMilne can you test and confirm?