Updated benchmarks.md.
Removed sundown, because the reading was anomalous.

This commit in hoedown caused the speed difference btw
sundown and hoedown that I was measuring before (on 32 bit
machines):

https://github.com/hoedown/hoedown/commit/ca829ff83580ed52cc56c09a67c80119026bae20

As Nick Wellnhofer explains: "The commit removes a rather arbitrary
limit of 16MB for buffers. Your benchmark input probably results in
an buffer larger than 16MB. It also seems that hoedown didn't check
error returns thoroughly at the time of the commit. This basically means
that large input files ould produce any kind of random behavior before
that commit, and that any benchmark that results in a too large buffer
can't be relied on."