Preliminary changes for new tab handling.
We no longer preprocess tabs to spaces before parsing.

Instead, we keep track of both the byte offset and
the (virtual) column as we parse block starts.

This allows us to handle tabs without converting
to spaces first.  Tabs are left as tabs in the output.

Added `column` and `first_nonspace_column` fields to `parser`.

Added utility function to advance the offset, computing
the virtual column too.

Note that we don't need to deal with UTF-8 here at all.
Only ASCII occurs in block starts.

Significant performance improvement due to the fact that
we're not doing UTF-8 validation -- though we might want
to add that back in.