Simpler approach for entity lookup.
We dispense with the hashes and just do string comparsions.
Since the array is in order, we can search intelligently
and should never need to do more than 8 or so comparisons.

This reduces binary size even further, at a small cost
in performance.  (This shouldn't matter too much, as
it's only detectable in really entity-heavy sources.)