Added RAW_BLOCK and RAW_INLINE node types.
These are passed through verbatim by all writers, with no
escaping.

They are never generated by the parser, and do not correspond
to CommonMark elements.  They are designed to be inserted by
filters that postprocess the AST.  For example, a filter might
convert specially marked code blocks to svg diagrams in HTML
and tikz diagrams in LaTeX, passing these through to the renderer
as a RAW_BLOCK.