render:  added begin_content field.
This is like `begin_line` except that it doesn't trigger
production of the prefix.  So it can be set after an initial
prefix (say `> `) is printed by the renderer, and consulted
in determining whether to escape content that has a special
meaning at the beginning of a line.

Used in the commonmark renderer.