roundtrip tests:  remove spurious failures.
In the commonmark writer we separate lists, and lists and
indented code, using a dummy HTML comment rather than two
blank lines (this is more portable).

So in evaluating the round-trip tests, we now strip out
these comments.

We also normalize HTML to avoid issues having to do with
line breaks.