Ported robinst's changes to link parsing.
See https://github.com/jgm/commonmark.js/pull/101

This uses a separate stack for brackets, instead of
putting them on the delimiter stack.  This avoids the
need for looking through the delimiter stack for the next
bracket.

It also avoids a shortcut reference lookup when the reference
text contains brackets.

The change dramatically improved performance on the nested links
pathological test for commonmark.js.  It has a smaller but measurable
effect here.