Tagfilter extension
When we encounter a tag that causes an HTML 5 parser's content model
flag [1] to be changed to RCDATA, CDATA or RAWTEXT [2] [3], we escape
the tag by replacing its opening "<" with "&lt;".  This causes the tag
to appear verbatim in the page it's placed on.

We do this to prevent users breaking the page content, where the parser
would not interpret further tags as inserted by cmark as HTML until a
matching close tag was hit.  (Such a closing tag could exist if a user
entered it themselves, but it'd cause all cmark-generated markup in
between to be rendered raw, and is unlikely to be desireable behaviour.)

[1] https://www.w3.org/TR/2009/WD-html5-20090423/syntax.html#tokenization
[2] https://www.w3.org/TR/2009/WD-html5-20090212/serializing-html-fragments.html#parsing-html-fragments
[3] https://github.com/google/gumbo-parser/blob/aa91b27b02c0c80c482e24348a457ed7c3c088e0/src/parser.c#L4023-L4053