test: Add test case for pathological collisions
As reported by @mity. The reference map implementation in cmark can be
trivially DDoSed by crafting a document where all link references have
names that collide.

Generating a list of colliding reference names is trivial, since the
hash function we use is not cryptographically safe, and our hash table
always has a fixed amount of buckets.

You can see an example on how to generate such documents in the
`hash_collisions` helper in `pathological_tests.py`: a document with
50000 link references, all colliding on the same bucket, is generated.

Parsing such Markdown document takes more than 5 minutes in the current
version of CMark. In order to allow the test suite to finish, and to
allow us to test other pathological inputs in the future, the
pathological test runner has been modified to use Python's `multiprocess`
module and spawn a worker to run each test.

With a worker running on a separate process, we can indeed see that that
the generated document does not finish parsing in a reasonable amount of
time. The timeout for all pathological tests has been set to 5 seconds,
which may need tuning for slow CI environments in the future.