Don't use CMAKE_INSTALL_LIBDIR to create libcmark.pc.
For some reason this wasn't getting set in processing
libcmark.pc.in, and we were getting the wrong entry in libcmark.pc.
(See #236)

The new approach sets an internal libdir variable to
lib${LIB_SUFFIX}.  This variable is used both to set the
install destination and in the libcmark.pc.in template.

Closes #236.

However, I'd welcome comments from @juhp who originally
added CMAKE_INSTALL_LIBDIR in #185.  I think that the new
system should work fine with Fedora, since LIB_SUFFIX will
be set appropriately, but some testing is in order.