Link executable with static or shared library
If CMARK_STATIC is on (default), link the executable with the static
library. This produces exactly the same result as compiling the library
sources again and linking with the object files.

If CMARK_STATIC is off, link the executable with the shared library.
This wasn't supported before and should be the preferred way to
package cmark on Linux distros.

Building only a shared library and a statically linked executable
isn't supported anymore but this doesn't seem useful.