assorted fixes to the build system
Remove the use of `CMARK_INLINE` (as is being considered upstream) as
this workaround is no longer needed. It was to support VS2013 which has
been obsoleted for some time now.

Restructure the CMake modules to match expectations for CMake.

Avoid defining `likely` and `unlikely` in global scope. While doing so,
prefer the modern `__has_builtin` for the check (with fallbacks for
pre-GCC 4.10, clang 3.4, and MSVC). This largely shouldn't matter in
practice as modern branch predictors do well without the hint. Sink the
declaration into source files as the macros are not namespaced.

Remove the now unnecessary `config.h`.

Clean up the builds to only support a single build type which is
dependent on the standard `BUILD_SHARED_LIBS`.

Simplify and normalise the install rules to use the `GNUInstallDirs`
parameters.